requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
DateTime==4.3
idna==3.3
numexpr==2.7.3
numpy==1.21.2
pandas==1.3.3
pkg-resources==0.0.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
RPi.GPIO==0.7.0
six==1.16.0
spidev==3.5
twilio==7.2.0
urllib3==1.26.7
zope.interface==5.4.0